Fxied Rolling Corr, Added Run with "inverted' quotes vs "reported' quotes
SW was copying rolling VAR instead of Corr....
The values are very similar to those of BBRG, but the series is more volatile. It seems that BBRG does some smoothing in the correlation chart?  The point being that the Corr in the plot is the same as in the Descriptive stats and we already saw that this was almost the same result as in bbrg...

Ran an alternative input se (_Alt_FX), with all the currencies quoted in local currency (indirect) and inverted them (1/x) to get the direct (US$) quotes. we can compare the results...  Putting the quotes in US$ (direct) makes the most sense in the log returns chart and the corr matrix.